requirements.txt
einops==0.8.0
numpy==2.0.1
opencv_python==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
scipy==1.14.1

